pyapp/pyproject.toml
[project]
name = "pyapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "httpx[socks]>=0.28.1",
    "openai>=2.11.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "typer>=0.20.0",
    "uvicorn>=0.38.0",
]
